- removing BinaryClassification example for the article simplicity
