+++ csharp-advanced-topics/MachineLearningInCsharp/MachineLearningInCsharp/MachineLearningInCsharp.csproj
@@ -16,7 +16,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="binaryCreditClassificationModel.zip">
+    <None Update="trainedModel.zip">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
   </ItemGroup>
