+++ csharp-advanced-topics/MachineLearningInCsharp/MachineLearningInCsharpEngine/DataModels/MulticlassClassification/ModelBuilder.cs
@@ -33,6 +33,12 @@ public class ModelBuilder
         predictionEngine = mlContext.Model.CreatePredictionEngine<ModelInput, ModelOutput>(mlModel);
     }
 
+    public void LoadModel(string path)
+    {
+        mlModel = mlContext.Model.Load(path, out _);
+        predictionEngine = mlContext.Model.CreatePredictionEngine<ModelInput, ModelOutput>(mlModel);
+    }
+
     public IEstimator<ITransformer> ProcessData()
     {
         var pipeline = mlContext.Transforms.Conversion
