+++ csharp-advanced-topics/MachineLearningInCsharp/MachineLearningInCsharpEngine/DataModels/MulticlassClassification/ModelOutput.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.ML.Data;
-
 namespace MachineLearningInCsharpEngine.DataModels.MulticlassClassification;
 
 public class ModelOutput
